renamed VerifiedLocationData
