--- Assets/Radar/Scripts/Models/VerifiedLocationResult.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f8929c68b15705c48a565132f68cb260
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
